changed starts and average rating to double and rounded to 0.1

+++ src/main/java/com/leshen/LetsEatRestaurantAPI/Contract/RestaurantListDto.java
@@ -17,5 +17,5 @@ public class RestaurantListDto {
     private List<TableDto> tables;
     @Setter
     @Getter
-    private Short stars;
+    private double stars;
 }
